Change ValueError

+++ frontend/src/services/Api.js
@@ -2,9 +2,12 @@ import axios from 'axios';
 import _ from 'lodash';
 import LocalStorageKeys from '../config/LocalStorageKeys';
 
-const URL = 'http://127.0.0.1:8000/api/';
+// const URL = 'http://127.0.0.1:8000/api/';
+const URL = '/api/';
 
 class Api {
+  cache = {};
+
   constructor(baseUrl) {
     this.api = axios.create({
       baseURL: baseUrl,
@@ -14,7 +17,7 @@ class Api {
     this.api.interceptors.response.use(
       response => response.data,
       error => {
-        if (error.response.status === 401) {
+        if (error.response && (error.response.status === 401)) {
           localStorage.clear();
 
           // Hacky-ish and ugly but works and quick
@@ -38,7 +41,7 @@ class Api {
 
   async login(username, password) {
     try {
-      const res = await this.api.post('login', {
+      const res = await this.api.post('login/', {
         username,
         password
       });
@@ -54,86 +57,68 @@ class Api {
   }
 
   async getApartment() {
-    try {
-      const res = await this.api.get('apartments');
-      return res[0];
-    } catch (e) {
-      throw e;
+    if (!this.cache.apartment) {
+      const res = await this.api.get('apartments/');
+      if (!res.length) throw new Error('No apartment registered to you.');
+      this.cache.apartment = res[0]; // eslint-disable-line
     }
+    return this.cache.apartment;
   }
 
   async getAvailableServices() {
-    return this.api.get('available-services');
+    return this.api.get('available-services/');
   }
 
-  async getSubscribedServices() {
-    try {
-      const res = await this.api.get('subscriptions');
-      localStorage.setItem(
-        LocalStorageKeys.SUBSCRIBED_SERVICES,
-        JSON.stringify(res)
-      );
+  async  getServiceSubscriptions() {
+    const res = await this.api.get('subscriptions/');
+    localStorage.setItem(
+      LocalStorageKeys.SUBSCRIBED_SERVICES,
+      JSON.stringify(res)
+    );
 
-      return res;
-    } catch (e) {
-      throw e;
-    }
+    return res;
   }
 
-  addSubscribedService(id) {
+  async getCustomServiceSubscriptions() {
+    return this.api.get('custom_subscriptions/');
+  }
+
+  addSubscribedService(id, apsenAttrIds, includeHistory) {
     return this.api.post('subscriptions/', {
-      service: id
-    });
+      service: id,
+      attributes: apsenAttrIds,
+      include_history: includeHistory
+    }, {timeout: 20000});
   }
 
   deleteSubscribedService(id) {
-    return this.api.delete(`subscriptions/${id}`);
+    return this.api.delete(`subscriptions/${id}/`);
   }
 
-  async getApartmentSensors() {
-    try {
-      const res = await this.api.get('apartmentsensors');
-
-      return _.reduce(
-        res,
-        (sensors, a) => {
-          const s = _.map(
-            a.apartment_sensor_values,
-            ({
-              description,
-              uri,
-              ui_type: uiType,
-              value,
-              updated_at: updatedAt
-            }) => ({
-              id: `${a.id}-${description.substr(0, 5)}`,
-              name: a.sensor.description,
-              identifier: a.identifier,
+  async getSensorValues() {
+    const apartment = await this.getApartment();
+    return _.flatten(
+      apartment.apartment_sensors.map(apsen =>
+        _.filter(apsen.attributes, 'ui_type').map(
+          ({
+            description,
+            uri,
+            ui_type: uiType,
+            value,
+            updated_at: updatedAt
+          }) => ({
+              id: `${apsen.id}-${description.substr(0, 5)}`,
+              name: apsen.sensor.description,
+              identifier: apsen.identifier,
               uri,
               description,
               uiType,
               value,
               updatedAt
             })
-          );
-
-          return _.concat(sensors, s);
-        },
-        []
-      );
-    } catch (e) {
-      throw e;
-    }
-  }
-
-  revokeApartment() {
-    const currentUser = JSON.parse(
-      localStorage.getItem(LocalStorageKeys.CURRENT_USER)
+        )
+      )
     );
-    const ID = currentUser && currentUser.id;
-    if (ID) return this.api.delete(`users/${ID}`);
-
-    throw new Error('User not logged in!');
   }
 }
